Refactor: Split `utils.dart` into multiple files
The `mapper/lib/src/utils.dart` file contained multiple classes and utility functions with different responsibilities. This made the code harder to navigate and maintain.

This commit refactors `utils.dart` by splitting its contents into three new files:
- `globals.dart`: Contains global utility functions `typeOf` and `kIsWeb`.
- `json_map.dart`: Contains the `JsonMap` class for traversing JSON objects.
- `class_info.dart`: Contains the `ClassInfo` class for reflection-based class analysis.

All import and export statements throughout the project have been updated to reflect this new file structure. The original `utils.dart` file has been removed.
