+++ mapper/lib/dart_json_mapper.dart
@@ -1,4 +1,6 @@
 export './src/errors.dart';
 export './src/mapper.dart';
 export './src/model/index.dart';
-export './src/utils.dart';
+export './src/globals.dart';
+export './src/json_map.dart';
+export './src/class_info.dart';
