+++ mapper/lib/src/globals.dart
@@ -1,0 +1,3 @@
+Type typeOf<T>() => T;
+
+const kIsWeb = identical(0, 0.0);
