+++ mapper/lib/src/model/type_info.dart
@@ -1,7 +1,8 @@
 import 'dart:collection';
 
+import '../class_info.dart';
 import './value_decorators.dart';
-import '../utils.dart';
+import '../globals.dart';
 
 /// Provides enhanced type information based on `Type.toString()` value
 class TypeInfo {
